fix: moved create task to dashboard
